[KFPv2_hell] google-cloud/Vertex_AI/Models - Added workaround_for_buggy_KFPv2_compiler versions for the Upload_moden and Deploy_model components
The KFPv2 compiler seems to have a bug, so it crashes when component uses non-primitive input types in conjunction with the inputValue placeholder.
These workarounds temporary fix the issue by erasing the input types and replacing them with strings.
This is not maintainable in the long run since input and output typing and type checking was one of KFP's advertized features.
The KFPv2 compiler bug needs to be fixed.