fix: Pandas - Fill_all_missing_values - Fixed filling the values for different column types
For CSV, all columns values are read as string, so the component fills the missing values in the original string form.
For Parquet, the columns have different types, so we need to correctly convert the replacement value.

In both cases the `replacement_type_name` parameter is not needed, so I've removed it.